move sanitize function to search-api.js
